feat: profileState 파일 분리 및 일정/테스트결과 id 앱유저 모델에 추가 메서드
